Modified InfoPanel_FaceCamera.cs to slerp the rotation if the SlerpRotate boolean is True. The speed at which it does so can be modified with SlerpRotateSpeed.
To avoid weird behaviour with the sky panels, before slerping it checks to see if the position this frame is different to the previous one (ie, if the tooltip has moved to a new point of interest). If it has, it performs a regular LookAt before slerping. Without this, the rotation is messed up upon moving until it slowly slerps into place, so this avoids that.