Modified the tooltip and wartable panel prefabs to include a label for the sprite image. To fit it in the panel had to be made slightly taller (200 -> 225). Other components on the tables were moved to compensate, and the description area was made a little larger. NOTE: For some reason, updating these prefabs did not update the ones in any existing scenes, so I had to remove the old ones from the scene and re-add them. I did this in ChrisWorkingScene (a duplicate of the latest client scene) - No others have been modified so the panels probably will look wrong in other scenes.
Wired up the close button (mostly) on the sky panel. Currently it just hides the panel, but this has the side effect of you being unable to reactivate the current node until a new one is activated, meaning that once the sky panel is gone, its gone for that node and can't be brought back without activating another first, so that will have to be dealt with.

Modified the sky panel rotation script to only rotate when not looking directly at the panel or any of its child objects, this should make it less annoying to activate the buttons on the panel.

Added the offset values for the POIs - These can be changed and need to be tested on a phone, but they work for now.
Values are as follows: 
	- POI_CentaurusA
		Y Offset from 0 -> 50 (X and Z unchanged)
	- POI_LargeMagCloud
		X offset from 0 -> 60
		Y Offset from 0 -> 70
		Z offset from 0 -> -53
	- POI_SupernovaRemnant
		Y Offset from 0 -> 60 (X and Z unchanged)
	- POI_VelaNebula
		Y Offset from 0 -> 150 (X and Z unchanged)
	- POI_VelaPulsar
		Y Offset from 0 -> -35 (X and Z unchanged)