Finished Subtitles, added photographs for Italian voice overs
Subtitles for all voice lines are implemented in Italian now. Ideally I would rework the subtitle system, but given only 3 lines are subtitled it isn't a major concern.

Also implemented photos for the Italian voice actors on the settings menu and fixed some missing Italian voice lines (the preview lines).

All that is missing now are a couple of lines of translation that I am waiting to be sent through - Once those are done, the Italian translation will be completed.